+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 - Colt King [ColtTheKing]
 
 # Developer Changelog
+### [0.0.9] 2017/10/18
+==================
+#### New Features
+- Updated Ticket class to draw text
+#### Changed Features
+- Cleaned up Burger class
+
 ### [0.0.8] 2017/10/17
 ==================
 #### Changed Features